+++ aidial_adapter_anthropic/_utils/json.py
@@ -13,10 +13,6 @@ from anthropic import Omit
 from pydantic import BaseModel
 
 
-def remove_nones(d: dict) -> dict:
-    return {k: v for k, v in d.items() if v is not None}
-
-
 def json_dumps_short(
     obj: Any, *, string_limit: int = 100, list_len_limit: int = 10, **kwargs
 ) -> str:
