+++ aidial_adapter_anthropic/_utils/pydantic.py
@@ -4,8 +4,3 @@ from pydantic import BaseModel
 class ExtraForbidModel(BaseModel):
     class Config:
         extra = "forbid"
-
-
-class ExtraAllowModel(BaseModel):
-    class Config:
-        extra = "allow"
