+++ aidial_adapter_anthropic/dial/attachments.py
@@ -20,8 +20,8 @@ from aidial_sdk.chat_completion import (
 )
 from pydantic import BaseModel
 
+from aidial_adapter_anthropic._utils.list import aiter_to_list
 from aidial_adapter_anthropic._utils.resource import Resource
-from aidial_adapter_anthropic._utils.stream import aiter_to_list
 from aidial_adapter_anthropic.adapter.errors import UserError, ValidationError
 from aidial_adapter_anthropic.dial._message import BaseMessage, SystemMessage
 from aidial_adapter_anthropic.dial.resource import (
